dialogflow-cx/vpc-sc-auth-server/server/requirements.txt
#!/bin/bash
# Copyright 2022 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#      http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
# shellcheck disable=SC1000-SC9999

Flask==2.0.2
types-Flask==1.1.6
Flask-Logging==0.1.3
gunicorn==20.1.0
requests==2.28.1
google-auth==2.14.1
google-cloud-secret-manager==2.12.3
google-cloud-storage==2.5.0
google-crc32c==1.3.0
pycryptodome==3.15.0